+++ src/AVR_Slow_PWM.h
@@ -18,11 +18,12 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.0.0
+  Version: 1.1.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.0.0   K.Hoang      27/09/2021 Initial coding for AVR-based boards  (UNO, Nano, Mega, 32U4, 16U4, etc. )
+  1.1.0   K Hoang      10/11/2021 Add functions to modify PWM settings on-the-fly
 *****************************************************************************************************************************/
 
 #pragma once
@@ -99,7 +100,7 @@
 #endif
 
 #ifndef AVR_SLOW_PWM_VERSION
-  #define AVR_SLOW_PWM_VERSION       F("AVR_Slow_PWM v1.0.0")
+  #define AVR_SLOW_PWM_VERSION       F("AVR_Slow_PWM v1.1.0")
 #endif
 
 #ifndef _PWM_LOGLEVEL_