+++ src/AVR_Slow_PWM.hpp
@@ -18,11 +18,12 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.0.0
+  Version: 1.1.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.0.0   K.Hoang      27/09/2021 Initial coding for AVR-based boards  (UNO, Nano, Mega, 32U4, 16U4, etc. )
+  1.1.0   K Hoang      10/11/2021 Add functions to modify PWM settings on-the-fly
 ****************************************************************************************************************************/
 
 #pragma once