+++ src_cpp/AVR_Slow_PWM_ISR.h
@@ -18,11 +18,12 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.0.0
+  Version: 1.1.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.0.0   K.Hoang      27/09/2021 Initial coding for AVR-based boards  (UNO, Nano, Mega, 32U4, 16U4, etc. )
+  1.1.0   K Hoang      10/11/2021 Add functions to modify PWM settings on-the-fly
 *****************************************************************************************************************************/
 
 #pragma once
@@ -99,7 +100,7 @@
 #endif
 
 #ifndef AVR_SLOW_PWM_VERSION
-  #define AVR_SLOW_PWM_VERSION       F("AVR_Slow_PWM v1.0.0")
+  #define AVR_SLOW_PWM_VERSION       F("AVR_Slow_PWM v1.1.0")
 #endif
 
 #ifndef _PWM_LOGLEVEL_
@@ -151,42 +152,71 @@ class AVR_Slow_PWM_ISR
     
     //////////////////////////////////////////////////////////////////
     // PWM
-    void setPWM(uint32_t pin, uint32_t frequency, uint32_t dutycycle, timer_callback StartCallback = nullptr,
+    // Return the channelNum if OK, -1 if error
+    int setPWM(uint32_t pin, double frequency, uint32_t dutycycle, timer_callback StartCallback = nullptr, 
                 timer_callback StopCallback = nullptr)
     {
       uint32_t period = 0;
       
-      if ( ( frequency != 0 ) && ( frequency <= 1000 ) )
+     if ( ( frequency != 0 ) && ( frequency <= 1000 ) )
       {
 #if USING_MICROS_RESOLUTION
       // period in us
-      period = 1000000 / frequency;
+      period = 1000000.0f / frequency;
 #else    
       // period in ms
-      period = 1000 / frequency;
+      period = 1000.0f / frequency;
 #endif
       }
       else
       {       
-        PWM_LOGERROR(F("Error: Invalid frequency, max is 500Hz"));
+        PWM_LOGERROR("Error: Invalid frequency, max is 1000Hz");
+        
+        return -1;
       }
       
-      setupPWMChannel(pin, period, dutycycle, (void *) StartCallback, (void *) StopCallback);     
+      return setupPWMChannel(pin, period, dutycycle, (void *) StartCallback, (void *) StopCallback);   
     }
 
-#if USING_MICROS_RESOLUTION
-    //period in us
-    void setPWM_Period(uint32_t pin, uint32_t period, uint32_t dutycycle, timer_callback StartCallback = nullptr,
-                       timer_callback StopCallback = nullptr)
-#else    
-    // PWM
-    //period in ms
-    void setPWM_Period(uint32_t pin, uint32_t period, uint32_t dutycycle, timer_callback StartCallback = nullptr,
-                       timer_callback StopCallback = nullptr)
-#endif    
+    // period in us
+    // Return the channelNum if OK, -1 if error
+    int setPWM_Period(uint32_t pin, uint32_t period, uint32_t dutycycle, timer_callback StartCallback = nullptr,
+                       timer_callback StopCallback = nullptr)  
     {     
-      setupPWMChannel(pin, period, dutycycle, (void *) StartCallback, (void *) StopCallback);       
-    }    
+      return setupPWMChannel(pin, period, dutycycle, (void *) StartCallback, (void *) StopCallback);      
+    } 
+    
+    //////////////////////////////////////////////////////////////////
+    
+    // low level function to modify a PWM channel
+    // returns the true on success or false on failure
+    bool modifyPWMChannel(unsigned channelNum, uint32_t pin, double frequency, uint32_t dutycycle)
+    {
+      uint32_t period = 0;
+      
+      if ( ( frequency > 0 ) && ( frequency <= 1000 ) )
+      {
+#if USING_MICROS_RESOLUTION
+      // period in us
+      period = 1000000.0f / frequency;
+#else    
+      // period in ms
+      period = 1000.0f / frequency;
+#endif
+      }
+      else
+      {       
+        PWM_LOGERROR("Error: Invalid frequency, max is 1000Hz");
+        return false;
+      }
+      
+      return modifyPWMChannel_Period(channelNum, pin, period, dutycycle);
+    }
+    
+    //////////////////////////////////////////////////////////////////
+    
+    //period in us
+    bool modifyPWMChannel_Period(unsigned channelNum, uint32_t pin, uint32_t period, uint32_t dutycycle);
     
     //////////////////////////////////////////////////////////////////
 
@@ -261,7 +291,6 @@ class AVR_Slow_PWM_ISR
     // actual number of PWM channels in use (-1 means uninitialized)
     volatile int numChannels;
 };
-
 
 
 #endif    // AVR_SLOW_PWM_ISR_H