+++ src_h/AVR_Slow_PWM.hpp
@@ -18,11 +18,12 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.0.0
+  Version: 1.1.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.0.0   K.Hoang      27/09/2021 Initial coding for AVR-based boards  (UNO, Nano, Mega, 32U4, 16U4, etc. )
+  1.1.0   K Hoang      10/11/2021 Add functions to modify PWM settings on-the-fly
 ****************************************************************************************************************************/
 
 #pragma once
@@ -30,8 +31,8 @@
 #ifndef AVR_SLOW_PWM_HPP
 #define AVR_SLOW_PWM_HPP
 
-#ifndef TIMER_INTERRUPT_DEBUG
-  #define TIMER_INTERRUPT_DEBUG      0
+#ifndef _PWM_LOGLEVEL_
+  #define _PWM_LOGLEVEL_      0
 #endif
 
 void TimerInterrupt::init(int8_t timer)