+++ changelog.md
@@ -12,12 +12,18 @@
 ## Table of Contents
 
 * [Changelog](#changelog)
+  * [Releases v1.1.0](#releases-v110)
   * [Initial Releases v1.0.0](#Initial-Releases-v100)
 
 ---
 ---
 
 ## Changelog
+
+### Releases v1.1.0
+
+1. Add functions to modify PWM settings on-the-fly
+2. Add example to demo how to modify PWM settings on-the-fly
 
 ### Initial Releases v1.0.0